Rename `rustc_driver_impl/session_diagnostics.rs` to `diagnostics.rs`
