+++ compiler/rustc_driver_impl/src/diagnostics.rs
